requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
google-generativeai==0.3.2
pillow>=9.0.0
aiofiles==23.2.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
# Email dependencies for OTP verification
jinja2==3.1.2
# Production dependencies
gunicorn==21.2.0
alembic==1.13.1
# Monitoring and logging
structlog==23.2.0
# Security
cryptography==41.0.8 